+++ tests/package.lisp
@@ -196,6 +196,8 @@
                 #:is-a/2
                 #:select/2
                 #:param/2
+                #:throw/1
+                #:catch/3
                 #:node-slot-value/3
                 #:retract/1
                 #:retract/3